+++ Marlin/src/lcd/menu/menu_bed_corners.cpp
@@ -28,7 +28,7 @@
 
 #if BOTH(HAS_LCD_MENU, LEVEL_BED_CORNERS)
 
-#include "menu.h"
+#include "menu_item.h"
 #include "../../module/motion.h"
 #include "../../module/planner.h"
