+++ Marlin/src/lcd/menu/menu_job_recovery.cpp
@@ -28,7 +28,7 @@
 
 #if BOTH(HAS_LCD_MENU, POWER_LOSS_RECOVERY)
 
-#include "menu.h"
+#include "menu_item.h"
 #include "../../gcode/queue.h"
 #include "../../sd/cardreader.h"
 #include "../../feature/powerloss.h"
