+++ Marlin/src/lcd/menu/menu_led.cpp
@@ -28,7 +28,7 @@
 
 #if HAS_LCD_MENU && EITHER(LED_CONTROL_MENU, CASE_LIGHT_MENU)
 
-#include "menu.h"
+#include "menu_item.h"
 
 #if ENABLED(LED_CONTROL_MENU)
   #include "../../feature/leds/leds.h"
