+++ Marlin/src/lcd/menu/menu_motion.cpp
@@ -28,7 +28,7 @@
 
 #if HAS_LCD_MENU
 
-#include "menu.h"
+#include "menu_item.h"
 #include "menu_addon.h"
 
 #include "../../module/motion.h"
