+++ Marlin/src/lcd/tft/touch.cpp
@@ -23,8 +23,9 @@
 
 #include "touch.h"
 
-#include "../ultralcd.h"
-#include "../menu/menu.h"
+#include "../ultralcd.h"  // for ui methods
+#include "../menu/menu.h" // for touch_screen_calibration
+
 #include "../../module/temperature.h"
 #include "../../module/planner.h"
